Refactored auth filters
