+++ src/Tests/Filters/MockFilter.php
@@ -7,7 +7,7 @@ use Lightpack\Utils\Url;
 
 class MockFilter implements IFilter
 {
-    public function before(Request $request)
+    public function before(Request $request, array $params = [])
     {
         if($request->isPost()) {
             $_POST['framework'] = 'Lightpack';
@@ -18,7 +18,7 @@ class MockFilter implements IFilter
         }
     }
 
-    public function after(Request $request, Response $response): Response
+    public function after(Request $request, Response $response, array $params = []): Response
     {
         $response->setBody('hello');
         return $response;
